tickets/src/routes/__tests__/new.test.ts
import request from "supertest";
import { app } from "../../app";
import { Ticket } from "../../models/ticket";
import { natsWrapper } from "../../nats-wrapper";

it("tem uma rota escutando em /api/tickets para requesicoes posts", async () => {
  const response = await request(app).post("/api/tickets").send({});

  expect(response.status).not.toEqual(404);
});

it("so pode ser acessado se o usuario estiver autenticado", async () => {
  const response = await request(app).post("/api/tickets").send({}).expect(401);
});

it("retorna um status diferente de 401 se o usuario estiver autenticado", async () => {
  const response = await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({});

  expect(response.status).not.toEqual(401);
});

it("retorna um erro se o um titulo inválido é fornecido", async () => {
  await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({
      title: "",
      price: 10,
    })
    .expect(400);
});

it("retrona um erro se um preço inválido é fornecido", async () => {
  await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({
      title: "dsadada",
      price: -10,
    })
    .expect(400);

  await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({
      title: "dsadada",
    })
    .expect(400);
});

it("cria um ticket com valores válidos", async () => {
  let tickets = await Ticket.find({});
  expect(tickets.length).toEqual(0);
  const title = "asdada";
  await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({ title, price: 20 })
    .expect(201);

  tickets = await Ticket.find({});
  expect(tickets.length).toEqual(1);
  expect(tickets[0].price).toEqual(20);
  expect(tickets[0].title).toEqual(title);
});

it("publicando um event", async () => {
  const title = "asdada";
  await request(app)
    .post("/api/tickets")
    .set("Cookie", global.signin())
    .send({ title, price: 20 })
    .expect(201);

  expect(natsWrapper.client.publish).toHaveBeenCalled();
});
